#1988: decompose pkg/flowexport into manager/netflow/ipfix/transport
Split the single-package flow exporter into responsibility-scoped files
and dedup the UDP transport code that the NetFlow v9 and IPFIX exporters
were each carrying a private copy of. This is pure code motion plus the
extraction of two shared transport helpers; the public API is unchanged
and the encoding/config logic is byte-identical to before.

Layout (per AGY review 012 II.3):

  manager.go    Resolved ExportConfig/CollectorConfig/SamplingDir, the
                FlowRecord and SessionCloseData shapes shared by both
                encoders, the sampling scheduler (ShouldExport), and the
                BuildExportConfig / BuildIPFIXExportConfig /
                BuildSamplingZones resolvers. Moved verbatim from the
                old exporter.go (with BuildIPFIXExportConfig and the
                FlowRecord type pulled over from ipfix.go and
                netflow9.go respectively).
  netflow.go    Renamed from netflow9.go. NetFlow v9 template/record
                encoding (unchanged) plus the Exporter that drives it,
                moved over from exporter.go.
  ipfix.go      IPFIX (v10) template/record encoding (unchanged) plus
                the IPFIXExporter; the config resolver moved out to
                manager.go.
  transport.go  New. collectorConns owns the dial loop, the per-packet
                fan-out write, and teardown that were duplicated
                line-for-line in NewExporter/NewIPFIXExporter and their
                send paths. flowBatch owns the per-family (v4/v6) record
                accumulation and atomic drain. Both exporters now hold a
                *collectorConns and a flowBatch instead of open-coding
                conns/batchMu/batchV4/batchV6.

dialCollectors takes []CollectorConfig (not the whole ExportConfig) so
the shared helper does not copy the atomic sampleCounter — go vet on the
package is unchanged from master (the two pre-existing "passes lock by
value" warnings on NewExporter/NewIPFIXExporter remain; no new ones).

Validation: full `go test ./...` green; encode/config function bodies
verified byte-identical to origin/master via per-function diff; gofmt
and go vet clean (no new findings) on the changed files. The
pkg/flowexport README is updated with the new file map.
